Move store constructor
